fix: use fp16 dtype to match available BEN2-ONNX model file
The repo only has model_fp16.onnx, no fp32 variant.

+++ src/App.tsx
@@ -20,7 +20,7 @@ async function getRemover(onProgress: (p: number) => void) {
   if (pipelineInstance) return pipelineInstance;
   pipelineInstance = (await pipeline("background-removal", "onnx-community/BEN2-ONNX", {
     device: "webgpu",
-    dtype: "fp32",
+    dtype: "fp16",
     progress_callback: (info: { status: string; progress?: number }) => {
       if (info.status === "progress" && info.progress != null) {
         onProgress(Math.round(info.progress));
